fix: accordion variant size
