+++ packages/react/src/components/accordion/components/Accordion.stories.tsx
@@ -1,13 +1,16 @@
+import { Story } from '@storybook/react';
 import * as React from 'react';
-import { IconChevronUp } from '@aksara-ui/icons';
-import styled, { keyframes } from 'styled-components';
-import { UnstyledButton } from '../../button';
+import { Accordion, AccordionContent, AccordionHeader, AccordionItem } from '.';
 import { Box } from '../../../layout';
-import { Heading } from '../../../typography';
-import { Accordion, AccordionItem, AccordionHeader, AccordionTrigger, AccordionContent } from '.';
 
 export default {
   title: 'Core/Components/Accordion',
+  argTypes: {
+    size: {
+      options: ['md', 'lg'],
+      control: { type: 'radio' },
+    },
+  },
   parameters: {
     backgrounds: {
       default: 'dark',
@@ -20,114 +23,35 @@ export default {
     },
   },
 };
-const slideDown = keyframes`
-  from  { height: 0 }
-  to { height: var(--radix-accordion-content-height) },`;
 
-const slideUp = keyframes`
-  from { height: var(--radix-accordion-content-height) }
-  to { height: 0 }`;
-
-const BoxContent = styled(Box)`
-  overflow: hidden;
-  &[data-state='open'] {
-    animation-name: ${slideDown};
-    animation-duration: 300ms;
-    animation-timing-function: cubic-bezier(0.87, 0, 0.13, 1);
-  }
-  &[data-state='closed'] {
-    animation-name: ${slideUp};
-    animation-duration: 300ms;
-    animation-timing-function: cubic-bezier(0.87, 0, 0.13, 1);
-  }
-`;
-
-const AccordionChevron = styled(IconChevronUp)`
-  transition: transform 300ms;
-  [data-state='open'] & {
-    transform: rotate(180deg);
-  }
-`;
-
-export const Example = () => {
+export const Example: Story<{ size: 'md' | 'lg' }> = ({ size }) => {
   return (
     <Accordion type="single" defaultValue="item-1" collapsible>
       <Box width={300} backgroundColor={'greylight01'} p="md">
         <AccordionItem value="item-1">
           <Box>
-            <AccordionHeader>
-              <Heading as="h3" display={'flex'} scale={300}>
-                <AccordionTrigger>
-                  <UnstyledButton
-                    py="sm"
-                    px="xs"
-                    height={40}
-                    display={'flex'}
-                    flex={1}
-                    borderRadius={'lg'}
-                    justifyContent={'space-between'}
-                    alignItems={'center'}
-                    _hover={{ backgroundColor: 'greylight03' }}
-                    _focus={{
-                      boxShadow: `0 0 0 2px rgba(175, 214, 255, 0.7)`,
-                    }}
-                    _focusVisible={{
-                      outline: 'none',
-                    }}
-                  >
-                    Title
-                    <AccordionChevron size={16} />
-                  </UnstyledButton>
-                </AccordionTrigger>
-              </Heading>
-            </AccordionHeader>
+            <AccordionHeader size={size}>Title</AccordionHeader>
             <AccordionContent>
-              <BoxContent>
-                <Box py="sm" px="xs">
-                  This is content
-                </Box>
-              </BoxContent>
+              <Box py="sm" px="xs">
+                This is content
+              </Box>
             </AccordionContent>
           </Box>
         </AccordionItem>
         <AccordionItem value="item-2">
           <Box>
-            <AccordionHeader>
-              <Heading as="h3" display={'flex'} scale={300}>
-                <AccordionTrigger>
-                  <UnstyledButton
-                    py="sm"
-                    px="xs"
-                    height={40}
-                    display={'flex'}
-                    flex={1}
-                    borderRadius={'lg'}
-                    justifyContent={'space-between'}
-                    alignItems={'center'}
-                    _hover={{ backgroundColor: 'greylight03' }}
-                    _focus={{
-                      boxShadow: `0 0 0 2px rgba(175, 214, 255, 0.7)`,
-                    }}
-                    _focusVisible={{
-                      outline: 'none',
-                    }}
-                  >
-                    Title
-                    <AccordionChevron size={16} />
-                  </UnstyledButton>
-                </AccordionTrigger>
-              </Heading>
-            </AccordionHeader>
+            <AccordionHeader size={size}>Title</AccordionHeader>
             <AccordionContent>
-              <BoxContent>
-                <Box py="sm" px="xs">
-                  This is content
-                </Box>
-              </BoxContent>
+              <Box py="sm" px="xs">
+                This is content
+              </Box>
             </AccordionContent>
           </Box>
         </AccordionItem>
       </Box>
     </Accordion>
   );
+};
+Example.args = {
+  size: 'md',
 };
